+++ examples/cryptocurrency-advanced/frontend/src/components/Navbar.vue
@@ -2,7 +2,10 @@
   <nav class="navbar navbar-expand-lg navbar-light bg-light">
     <div class="container">
       <router-link :to="{ name: isAuthorized ? 'user' : 'home' }" class="navbar-brand">Cryptocurrency</router-link>
-      <div class="collapse navbar-collapse">
+      <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbar" aria-controls="navbar" aria-expanded="false" aria-label="Toggle navigation">
+        <span class="navbar-toggler-icon"></span>
+      </button>
+      <div class="collapse navbar-collapse" id="navbar">
         <ul class="navbar-nav">
           <li class="nav-item">
             <router-link :to="{ name: 'blockchain' }" class="nav-link">Blockchain</router-link>