+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 ### New features
 
+#### exonum
+
+- The `--allow-origin` parameter has been added to the `finalize` command.
+
 - IPv6 addressing is now supported. (#615)
 
 ### Bug fixes